requirements.txt
discord.py>=1.5
jishaku>=1.18